chore(dynamic-sampling): quiet redis noise in schedule_invalidate_project_config (#101331)
