+++ src/sentry/workflow_engine/utils/__init__.py
@@ -1,8 +1,6 @@
 __all__ = [
     "metrics_incr",
     "MetricTags",
-    "timeout_grouping_context",
 ]
 
-from .exception_grouping import timeout_grouping_context
 from .metrics import MetricTags, metrics_incr
