+++ tests/sentry/utils/test_exceptions.py
@@ -1,10 +1,12 @@
+from collections.abc import Callable
 from typing import Any
 from unittest.mock import Mock, patch
 
 from sentry.taskworker.state import CurrentTaskState
 from sentry.taskworker.workerchild import ProcessingDeadlineExceeded
-from sentry.workflow_engine.utils.exception_grouping import (
+from sentry.utils.exceptions import (
     exception_grouping_context,
+    set_sentry_exception_levels,
     timeout_grouping_context,
 )
 
@@ -34,7 +36,7 @@ class TestExceptionGroupingContext:
         )
 
         with patch(
-            "sentry.workflow_engine.utils.exception_grouping.current_task",
+            "sentry.utils.exceptions.current_task",
             return_value=mock_task_state,
         ):
             with patch("sentry_sdk.new_scope") as mock_scope:
@@ -85,7 +87,7 @@ class TestExceptionGroupingContext:
         )
 
         with patch(
-            "sentry.workflow_engine.utils.exception_grouping.current_task",
+            "sentry.utils.exceptions.current_task",
             return_value=mock_task_state,
         ):
             with patch("sentry_sdk.new_scope") as mock_scope:
@@ -156,7 +158,7 @@ class TestExceptionGroupingContext:
         )
 
         with patch(
-            "sentry.workflow_engine.utils.exception_grouping.current_task",
+            "sentry.utils.exceptions.current_task",
             return_value=mock_task_state,
         ):
             with patch("sentry_sdk.new_scope") as mock_scope:
@@ -193,10 +195,8 @@ class TestExceptionGroupingContext:
 
     def test_without_task_state(self) -> None:
         """Test that the context works when no task state is available."""
-        with patch(
-            "sentry.workflow_engine.utils.exception_grouping.current_task", return_value=None
-        ):
-            with patch("sentry.workflow_engine.utils.exception_grouping.logger") as mock_logger:
+        with patch("sentry.utils.exceptions.current_task", return_value=None):
+            with patch("sentry.utils.exceptions.logger") as mock_logger:
                 exception_mapping: dict[type[BaseException], str] = {
                     CustomError: "custom.error.fingerprint"
                 }
@@ -210,9 +210,7 @@ class TestExceptionGroupingContext:
     def test_context_manager_yields_correctly(self) -> None:
         """Test that the context manager yields correctly."""
         executed = False
-        with patch(
-            "sentry.workflow_engine.utils.exception_grouping.current_task", return_value=None
-        ):
+        with patch("sentry.utils.exceptions.current_task", return_value=None):
             exception_mapping: dict[type[BaseException], str] = {
                 CustomError: "custom.error.fingerprint"
             }
@@ -239,7 +237,7 @@ class TestExceptionGroupingContext:
             pass
 
         with patch(
-            "sentry.workflow_engine.utils.exception_grouping.current_task",
+            "sentry.utils.exceptions.current_task",
             return_value=mock_task_state,
         ):
             with patch("sentry_sdk.new_scope") as mock_scope:
@@ -289,7 +287,7 @@ class TestExceptionGroupingContext:
         )
 
         with patch(
-            "sentry.workflow_engine.utils.exception_grouping.current_task",
+            "sentry.utils.exceptions.current_task",
             return_value=mock_task_state,
         ):
             with patch("sentry_sdk.new_scope") as mock_scope:
@@ -339,7 +337,7 @@ class TestTimeoutGroupingContext:
         )
 
         with patch(
-            "sentry.workflow_engine.utils.exception_grouping.current_task",
+            "sentry.utils.exceptions.current_task",
             return_value=mock_task_state,
         ):
             with patch("sentry_sdk.new_scope") as mock_scope:
@@ -387,7 +385,7 @@ class TestTimeoutGroupingContext:
         )
 
         with patch(
-            "sentry.workflow_engine.utils.exception_grouping.current_task",
+            "sentry.utils.exceptions.current_task",
             return_value=mock_task_state,
         ):
             with patch("sentry_sdk.new_scope") as mock_scope:
@@ -418,3 +416,34 @@ class TestTimeoutGroupingContext:
                 # Event should be unchanged
                 assert result == {"original": "data"}
                 assert "fingerprint" not in result
+
+
+class TestSetSentryExceptionLevels:
+    def test_basic_functionality_minimal_mocking(self) -> None:
+        with patch("sentry_sdk.new_scope") as mock_scope:
+            mock_scope_instance = Mock()
+            mock_scope.return_value.__enter__ = Mock(return_value=mock_scope_instance)
+            mock_scope.return_value.__exit__ = Mock(return_value=None)
+
+            # Use a single-element list to capture the processor
+            captured_processors: list[Callable[[Any, Any], Any]] = []
+            mock_scope_instance.add_error_processor = captured_processors.append
+
+            # Use the context manager with exception type as key
+            with set_sentry_exception_levels({ValueError: "warning"}):
+                pass
+
+            # Basic validation that processor was captured
+            assert len(captured_processors) == 1
+            processor = captured_processors[0]
+
+            # Test that it correctly processes a ValueError
+            event = {"level": "error", "other_data": "preserved"}
+            exc = ValueError("test error")
+            exc_info = (ValueError, exc, None)
+
+            result = processor(event, exc_info)
+
+            # Verify the level was changed and other data preserved
+            assert result["level"] == "warning"
+            assert result["other_data"] == "preserved"
